Update patch set 2
Patch Set 2:

> Can somebody please tell me, what is left to be done to get this
 > change merged?

Well, it's waiting until someone who has the needed ACL and find time to review and verify it. It would help, if you could put verify label, or if you don't have the ACL, just comment that you have verified it, and (optional) provide the steps of your verification.

Patch-set: 2